+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## [UNRELEASED]
 
+- add .min() and .max() methods for ZArray
 - update build package to 3.0.2 and source_gen to 3.1.0
 
 ## 1.0.0 - 2025-08-19